Authentication

+++ code/js/api/lib/routes/employee-routes.js
@@ -22,8 +22,8 @@ router.get('/employees/logged', auth.optional(), (req, res, next) => {
 })
 
 router.get('/employees', auth.requested(), (req, res, next) => {
-    //if (!req.employee?.roles.includes('Manage Employees'))
-    //    next(new error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
+    if (!req.employee?.roles.includes('Manage Employees') && !req.employee?.roles.includes('Manage Section\'s Employees Roles'))
+        return next(error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
     service.getEmployees()
         .then(employees => res.send(
             new Entity(
@@ -32,23 +32,24 @@ router.get('/employees', auth.requested(), (req, res, next) => {
                 employeeSiren.getEmployeesLinks,
                 undefined,
                 [employeeSiren.addEmployeeAction()],
-                employeeSiren.setSubEntities(employees, emp => 
-                    [employeeSiren.updateEmployeeAction(emp._id),
-                    employeeSiren.deleteEmployeeAction(emp._id)])
+                employeeSiren.setSubEntities(
+                        employees.filter(employee => 
+                            employee.sections.some(section => 
+                                req.employee.sections.includes(section))), 
+                        emp => 
+                            [employeeSiren.updateEmployeeAction(emp._id),
+                            employeeSiren.deleteEmployeeAction(emp._id)])
             )))
         .catch(next)
 })
 
 router.post('/employees', auth.requested(), (req, res, next) => {
-    //if (!req.employee?.roles.includes('Manage Employees'))
-    //    next(new error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
+    if (!req.employee?.roles.includes('Manage Employees') && !req.employee?.roles.includes('Manage Section\'s Employees Roles'))
+        return next(error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
     const name = req.body.name
     const _id = req.body._id
     if(!name || !_id)
         throw error.CustomException('Missing required parameters', error.BAD_REQUEST)
-    const roles = req.body.roles
-    const sections = req.body.sections
-    const picture = req.body.picture
     model.employeeInputModel.validateAsync({_id, name})
         .then(employee => service.addEmployee(employee)
             .then(employee => res.status(201).send(
@@ -62,8 +63,8 @@ router.post('/employees', auth.requested(), (req, res, next) => {
 })
 
 router.patch('/employees/:employeeId', auth.requested(), (req, res, next) => {
-    //if (!req.employee?.roles.includes('Manage Employees'))
-    //    next(new error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
+    if (!req.employee?.roles.includes('Manage Employees') && !req.employee?.roles.includes('Manage Section\'s Employees Roles'))
+        return next(error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
     const _id = req.params.employeeId
     const name = req.body.name
     const roles = req.body.roles
@@ -85,8 +86,8 @@ router.patch('/employees/:employeeId', auth.requested(), (req, res, next) => {
 })
     
 router.delete('/employees/:employeeId', auth.requested(), (req, res, next) => {
-    //if (!req.employee?.roles.includes('Manage Employees'))
-    //    next(new error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
+    if (!req.employee?.roles.includes('Manage Employees') && !req.employee?.roles.includes('Manage Section\'s Employees Roles'))
+        return next(error.CustomException('You do not have permission to access this resource.', error.UNAUTHORIZED))
     const id = req.params.employeeId
     service.removeEmployee(id)
         .then(() => res.send(
